Implement WatchlistController, Watchlist.vue and associated models, migrations and policies.
This work is largely unfinished, but I am well on my way. I need to
figure out how the WatchlistItem components can tall the parent what
company it represents so that the parent can remove the correct child
component and perform associated ajax calls (creating watchlist,
adding/removing list items). Do I need to make a
WatchlistItemController? The logic behind storing the company, if it is
not already in the DB needs sorting out as well. Maybe look into
validation…